add redis lock for deleted sessions

+++ src/1_control/http/auth.controller.ts
@@ -17,6 +17,7 @@ import { AuthService } from '../../2_service/auth/auth.service';
 import { APP_DEFAULT_V1_PREFIX } from '../../5_shared/config/const/app.const';
 import { SWAGGER_TAGS } from '../../5_shared/config/const/swagger.const';
 import { ApiAbstractResponse } from '../../5_shared/decorators/api-abstract-response.decorator';
+import { GetAccessTokenFromRequest } from '../../5_shared/decorators/get-access-token-from-request';
 import { GetUserFromRequest } from '../../5_shared/decorators/get-user-from-request.decorator';
 import {
   GetUserAgentAndIp,
@@ -149,13 +150,14 @@ export class AuthController {
   @ApiAbstractResponse(SessionDto)
   @Delete('sessions/:sessionId')
   async deleteSession(
+    @GetAccessTokenFromRequest() accessToken: string,
     @Param('sessionId', ParseUUIDPipe) sessionId: string,
   ): Promise<void> {
     this.logger.info(
       `${this.deleteSession.name}, session id: ${JSON.stringify(sessionId)}`,
     );
 
-    await this.authService.deleteSession(sessionId);
+    await this.authService.deleteSession({ id: sessionId, accessToken });
   }
 
   @ApiOperation({
